+++ src/debug/ImGuiMeshWrapper.cpp
@@ -11,7 +11,7 @@ namespace ImGui {
 	{
 		static const std::string title = "Mesh ";
 		
-		ImGui::PushID(meshId);
+		ImGui::PushID(static_cast<int>(meshId));
 		if (ImGui::CollapsingHeader(mesh.GetName().empty() ? (title + std::to_string(meshId)).c_str() : mesh.GetName().c_str()))
 		{
 			ImGui::Indent();
